refactor: rename ProductSizeSelector component to match naming conventions
